+++ src/Rocks.Tests/Generators/HiddenMembersGeneratorTests.cs
@@ -6,6 +6,186 @@ namespace Rocks.Tests.Generators;
 public static class HiddenMembersGeneratorTests
 {
 	[Test]
+	public static async Task GenerateWithSubclassHidingAndReturnTypeIsDifferentAsync()
+	{
+		var code =
+			"""
+			using Rocks;
+
+			#nullable enable
+
+			namespace MockTests
+			{
+				public class Credentials { }
+
+				public class NewCredentials : Credentials { }
+			
+				public class BaseCopy
+				{
+					public virtual void Ok() { }
+					public virtual Credentials Copy() => new();
+				}
+
+				public class NewCopy : BaseCopy
+				{
+					public new NewCredentials Copy() => new();
+				}
+
+				public static class Test
+				{
+					public static void Generate()
+					{
+						var rock = Rock.Create<NewCopy>();
+					}
+				}
+			}
+			""";
+
+		var generatedCode =
+			"""
+			using Rocks.Extensions;
+			using System.Collections.Generic;
+			using System.Collections.Immutable;
+			#nullable enable
+			
+			namespace MockTests
+			{
+				internal static class CreateExpectationsOfNewCopyExtensions
+				{
+					internal static global::Rocks.Expectations.MethodExpectations<global::MockTests.NewCopy> Methods(this global::Rocks.Expectations.Expectations<global::MockTests.NewCopy> @self) =>
+						new(@self);
+					
+					internal static global::MockTests.NewCopy Instance(this global::Rocks.Expectations.Expectations<global::MockTests.NewCopy> @self)
+					{
+						if (!@self.WasInstanceInvoked)
+						{
+							@self.WasInstanceInvoked = true;
+							var @mock = new RockNewCopy(@self);
+							@self.MockType = @mock.GetType();
+							return @mock;
+						}
+						else
+						{
+							throw new global::Rocks.Exceptions.NewMockInstanceException("Can only create a new mock once.");
+						}
+					}
+					
+					private sealed class RockNewCopy
+						: global::MockTests.NewCopy
+					{
+						private readonly global::System.Collections.Generic.Dictionary<int, global::System.Collections.Generic.List<global::Rocks.HandlerInformation>> handlers;
+						
+						public RockNewCopy(global::Rocks.Expectations.Expectations<global::MockTests.NewCopy> @expectations)
+						{
+							this.handlers = @expectations.Handlers;
+						}
+						
+						[global::Rocks.MemberIdentifier(0, "bool Equals(object? @obj)")]
+						public override bool Equals(object? @obj)
+						{
+							if (this.handlers.TryGetValue(0, out var @methodHandlers))
+							{
+								foreach (var @methodHandler in @methodHandlers)
+								{
+									if (((global::Rocks.Argument<object?>)@methodHandler.Expectations[0]).IsValid(@obj))
+									{
+										@methodHandler.IncrementCallCount();
+										var @result = @methodHandler.Method is not null ?
+											((global::System.Func<object?, bool>)@methodHandler.Method)(@obj) :
+											((global::Rocks.HandlerInformation<bool>)@methodHandler).ReturnValue;
+										return @result!;
+									}
+								}
+								
+								throw new global::Rocks.Exceptions.ExpectationException("No handlers match for bool Equals(object? @obj)");
+							}
+							else
+							{
+								return base.Equals(@obj);
+							}
+						}
+						
+						[global::Rocks.MemberIdentifier(1, "int GetHashCode()")]
+						public override int GetHashCode()
+						{
+							if (this.handlers.TryGetValue(1, out var @methodHandlers))
+							{
+								var @methodHandler = @methodHandlers[0];
+								@methodHandler.IncrementCallCount();
+								var @result = @methodHandler.Method is not null ?
+									((global::System.Func<int>)@methodHandler.Method)() :
+									((global::Rocks.HandlerInformation<int>)@methodHandler).ReturnValue;
+								return @result!;
+							}
+							else
+							{
+								return base.GetHashCode();
+							}
+						}
+						
+						[global::Rocks.MemberIdentifier(2, "string? ToString()")]
+						public override string? ToString()
+						{
+							if (this.handlers.TryGetValue(2, out var @methodHandlers))
+							{
+								var @methodHandler = @methodHandlers[0];
+								@methodHandler.IncrementCallCount();
+								var @result = @methodHandler.Method is not null ?
+									((global::System.Func<string?>)@methodHandler.Method)() :
+									((global::Rocks.HandlerInformation<string?>)@methodHandler).ReturnValue;
+								return @result!;
+							}
+							else
+							{
+								return base.ToString();
+							}
+						}
+						
+						[global::Rocks.MemberIdentifier(3, "void Ok()")]
+						public override void Ok()
+						{
+							if (this.handlers.TryGetValue(3, out var @methodHandlers))
+							{
+								var @methodHandler = @methodHandlers[0];
+								@methodHandler.IncrementCallCount();
+								if (@methodHandler.Method is not null)
+								{
+									((global::System.Action)@methodHandler.Method)();
+								}
+							}
+							else
+							{
+								base.Ok();
+							}
+						}
+						
+					}
+				}
+				
+				internal static class MethodExpectationsOfNewCopyExtensions
+				{
+					internal static global::Rocks.MethodAdornments<global::MockTests.NewCopy, global::System.Func<object?, bool>, bool> Equals(this global::Rocks.Expectations.MethodExpectations<global::MockTests.NewCopy> @self, global::Rocks.Argument<object?> @obj)
+					{
+						global::System.ArgumentNullException.ThrowIfNull(@obj);
+						return new global::Rocks.MethodAdornments<global::MockTests.NewCopy, global::System.Func<object?, bool>, bool>(@self.Add<bool>(0, new global::System.Collections.Generic.List<global::Rocks.Argument>(1) { @obj }));
+					}
+					internal static global::Rocks.MethodAdornments<global::MockTests.NewCopy, global::System.Func<int>, int> GetHashCode(this global::Rocks.Expectations.MethodExpectations<global::MockTests.NewCopy> @self) =>
+						new global::Rocks.MethodAdornments<global::MockTests.NewCopy, global::System.Func<int>, int>(@self.Add<int>(1, new global::System.Collections.Generic.List<global::Rocks.Argument>()));
+					internal static global::Rocks.MethodAdornments<global::MockTests.NewCopy, global::System.Func<string?>, string?> ToString(this global::Rocks.Expectations.MethodExpectations<global::MockTests.NewCopy> @self) =>
+						new global::Rocks.MethodAdornments<global::MockTests.NewCopy, global::System.Func<string?>, string?>(@self.Add<string?>(2, new global::System.Collections.Generic.List<global::Rocks.Argument>()));
+					internal static global::Rocks.MethodAdornments<global::MockTests.NewCopy, global::System.Action> Ok(this global::Rocks.Expectations.MethodExpectations<global::MockTests.NewCopy> @self) =>
+						new global::Rocks.MethodAdornments<global::MockTests.NewCopy, global::System.Action>(@self.Add(3, new global::System.Collections.Generic.List<global::Rocks.Argument>()));
+				}
+			}
+			
+			""";
+
+		await TestAssistants.RunAsync<RockCreateGenerator>(code,
+			new[] { (typeof(RockCreateGenerator), "MockTests.NewCopy_Rock_Create.g.cs", generatedCode) },
+			Enumerable.Empty<DiagnosticResult>()).ConfigureAwait(false);
+	}
+
+	[Test]
 	public static async Task GenerateWithClassWhenMembersAreHiddenAsync()
 	{
 		var code =